+++ src/Grid.cpp
@@ -49,22 +49,22 @@ void Grid::printGrid(Evolve::TextureRenderer& renderer) {
     }
 }
 
-void Grid::addSnakeCell(const int row, const int column) {
+void Grid::addSnakeCell(const size_t row, const size_t column) {
     grid_[row][column] = SNAKE_SIGN;
 }
 
-void Grid::addFruitCell(const int row, const int column) {
+void Grid::addFruitCell(const size_t row, const size_t column) {
     grid_[row][column] = FRUIT_SIGN;
 }
 
-void Grid::addJackpotCells(const int row, const int column) {
+void Grid::addJackpotCells(const size_t row, const size_t column) {
     grid_[row][column] = JACKPOT_SIGN;
     grid_[row][column + 1] = JACKPOT_SIGN;
     grid_[row + 1][column] = JACKPOT_SIGN;
     grid_[row + 1][column + 1] = JACKPOT_SIGN;
 }
 
-void Grid::clearCell(const int row, const int column, bool isJackpot /*= false*/) {
+void Grid::clearCell(const size_t row, const size_t column, bool isJackpot /*= false*/) {
     grid_[row][column] = EMPTY_SIGN;
     
     if (isJackpot) {
@@ -74,18 +74,18 @@ void Grid::clearCell(const int row, const int column, bool isJackpot /*= false*/
     }
 }
 
-bool Grid::isEmptyCell(const int row, const int column) const {
+bool Grid::isEmptyCell(const size_t row, const size_t column) const {
     return grid_[row][column] == EMPTY_SIGN;
 }
 
-bool Grid::isSnakeCell(const int row, const int column) const {
+bool Grid::isSnakeCell(const size_t row, const size_t column) const {
     return grid_[row][column] == SNAKE_SIGN;
 }
 
-bool Grid::isFruitCell(const int row, const int column) const {
+bool Grid::isFruitCell(const size_t row, const size_t column) const {
     return grid_[row][column] == FRUIT_SIGN;
 }
 
-bool Grid::isJackpotCell(const int row, const int column) const {
+bool Grid::isJackpotCell(const size_t row, const size_t column) const {
     return grid_[row][column] == JACKPOT_SIGN;
 }
